pyproject.toml
[project]
name = "secretary"
version = "0.1.0"
description = "Telegram Secretary Bot"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.11",
    "beautifulsoup4>=4.12.3",
    "python-telegram-bot[callback-data]>=21.10",
    "wcpan-jav>=0.1.0",
    "wcpan-logging>=2.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.2",
    "wcpan-watchdog>=1.2.1",
]
